feat: make `LintGroup` a trait and delete `BevyLint`
`LintGroup` will now be implemented for a ZST struct in each lint group module. This trait will contain all of the information the old `LintGroup` struct did, and will also handle registration of those lints, passes, and itself.

Because each `LintGroup` now knows which lints are in it, individuals lints no longer need to track their associated `LintGroup`. This means that `BevyLint` is no longer needed, which is nice! (I have yet to update the macros, but that's next on my list.)